requirements.txt
pytorch==1.1.0
numpy==1.16.3
matplotlib==2.1.2
tqdm==4.26.0
recordtype
yamlordereddictloader
tabulate
